+++ .github/workflows/release.yml
@@ -1,0 +1,30 @@
+name: Release Packages
+
+on:
+  release:
+    types: [released]
+
+jobs:
+  publish-to-npm-registry:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+
+      - uses: actions/setup-node@v4
+        with:
+          node-version: '20.x'
+
+      - run: npm install
+
+      # build and package
+
+      - run: npm run build
+
+      - run: npm run package
+
+      - run: npx publib-npm
+        env:
+          NPM_TOKEN: ${{secrets.NPM_PUBLISH_TOKEN}}
+          NPM_ACCESS_LEVEL: public
