+++ lib/index.ts
@@ -1,1 +1,5 @@
 export * from "./oncoanalyser";
+export * from "./nextflow-config";
+export * from "./nextflow-task-environment";
+export * from "./nextflow-pipeline-environment";
+export * from "./oncoanalyser-job-definition";
