+++ lib/nextflow-pipeline-environment/index.ts
